Disable swipe gestures
jQuery Mobile's swipe events are not perfect.  It checks only the start
point and the end point of a gesture.  For example, if you drag down
then right like a letter "L", this gesture triggers a swiperight event.
As a result, swipe events are unexpectedly triggered by non-swipe
gestures.  It's very annoying.